last update on this branch since im gooing to india the next day :/
